requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.12.1
filelock==3.0.12
mysqlclient==2.0.1
protobuf==3.13.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.33
